hw_21

+++ application/views.py
@@ -1,4 +1,4 @@
-from django.shortcuts import render
+from django.shortcuts import render, redirect
 from django.http import HttpResponse
 from application.models import Shoppinglist, Item, Malllist, UserList
 
@@ -6,6 +6,8 @@ from application.models import Shoppinglist, Item, Malllist, UserList
 
 
 def index(request):
+    if not request.user.is_authenticated:
+        return redirect('/user/login')
     user_list = UserList.objects.filter(id=1).first()
     if request.method == 'POST':
         item_name = request.POST.get('item')
@@ -27,30 +29,26 @@ def index(request):
 
 
 def buy_item(request, item_id):
-    user_list = UserList.objects.filter(id=1).first()
+    if not request.user.is_authenticated:
+        return redirect('/user/login')
     if request.method == 'POST':
         item_id = request.POST.get('item')
         item_object = Item.objects.filter(pk=int(item_id)).first()
         shoppinglist_obj = Shoppinglist.objects.filter(item_id=item_object).first()
         shoppinglist_obj.status = 'bought'
         shoppinglist_obj.save()
-    result = Shoppinglist.objects.filter(list_id=user_list.list_id)
-    return render(request, 'buy_item.html',
-                  {'shoppinglist_data': result,
-                   'items': Item.objects.all().filter(shop_id__list_id=user_list.list_id)})
+    return redirect('/shoppinglist')
 
 
 def remove_item(request, item_id):
-    user_list = UserList.objects.filter(id=1).first()
+    if not request.user.is_authenticated:
+        return redirect('/user/login')
     if request.method == 'POST':
         item_id = request.POST.get('item')
         item_object = Item.objects.filter(pk=int(item_id)).first()
         item_obj = Item.objects.filter(id=item_object.id).first()
         item_obj.delete()
-    result = Shoppinglist.objects.filter(list_id=user_list.list_id)
-    return render(request, 'remove_item.html',
-                  {'shoppinglist_data': result,
-                   'items': Item.objects.all().filter(shop_id__list_id=user_list.list_id)})
+    return redirect('/shoppinglist')
 
 
 def index_user(request):
